Add files via upload
"Recursive Array Challenges"

Description:
This repository contains a collection of programming challenges focused on recursion with an emphasis on array manipulation. Whether you're a beginner looking to improve your recursion skills or an experienced coder looking for a refresher, these questions will help you explore the power of recursion in solving complex problems involving arrays.
Following are the questions in the attached file:
      1 - CHECK IF ARRAY IS SORTED OR NOT USING RECURSION
      2 - LINEAR SEARCH USING RECURSION
      3 - FINDING ALL OCCURRENCES OF THE TARGET ELEMENT IN ARRAY
                  -> USING STATIC ARRAYLIST
                  -> PASSING ARRAYLIST AS AN ARGUMENT 
                  -> MAKING ARRAYLIST INSIDE THE FUNCTION
      4 - FINDING TARGET ELEMENT IN A ROTATED ARRAY 

Feel free to fork this repository, submit your solutions, or use it as a practice resource to enhance your understanding of recursion and its applications in array-based problems. Happy coding!